Move a bunch more functions into the new module.
